# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shadow Garden API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  success: boolean;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface SendOtpBody {
  /** Phone number in international format (e.g. +2347012345678) */
  phone: string;
}

export interface SendOtpResponse {
  success: boolean;
  message: string;
}

export interface VerifyOtpBody {
  phone: string;
  code: string;
}

export interface UserProfile {
  id: string;
  name: string;
  phone?: string;
  level: number;
  xp: number;
  balance: number;
  bank: number;
  premium?: number;
  bio?: string;
  registeredAt?: number;
}

export interface AuthSession {
  success: boolean;
  token: string;
  user: UserProfile;
}

export interface RpgStats {
  class: string;
  hp: number;
  maxHp: number;
  attack: number;
  defense: number;
  speed: number;
  dungeonFloor: number;
}

export interface GuildMembership {
  id: string;
  name: string;
  level: number;
}

export interface UserStats {
  profile: UserProfile;
  rpg?: RpgStats;
  guild?: GuildMembership;
  rank: number;
  totalUsers: number;
  xpNeeded: number;
}

export interface InventoryItem {
  item: string;
  quantity: number;
  category: string;
}

export interface InventoryResponse {
  items: InventoryItem[];
}

export interface Achievement {
  id: number;
  name: string;
  description: string;
  icon: string;
  earnedAt: number;
}

export interface AchievementsResponse {
  achievements: Achievement[];
}

export interface ShopItem {
  id: number;
  name: string;
  description: string;
  price: number;
  effect: string;
  category: string;
}

export type ShopResponseCategoriesItem = {
  name: string;
  items: ShopItem[];
};

export interface ShopResponse {
  categories: ShopResponseCategoriesItem[];
}

export interface BuyItemBody {
  itemId: number;
  quantity?: number;
}

export interface BuyItemResponse {
  success: boolean;
  message: string;
  newBalance: number;
}

export interface Card {
  id: string;
  name: string;
  tier: string;
  series: string;
  description?: string;
  attack: number;
  defense: number;
  speed: number;
  totalCopies: number;
  ownerName?: string;
  ownerId?: string;
}

export interface CardsResponse {
  cards: Card[];
  total: number;
}

export interface UserCard {
  userCardId: number;
  card: Card;
  obtainedAt: number;
}

export interface UserCardsResponse {
  cards: UserCard[];
  total: number;
}

export interface WishlistBody {
  cardId: string;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  name: string;
  level: number;
  xp: number;
  guildName?: string | null;
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
}

export interface MyRankResponse {
  rank: number;
  total: number;
  entry: LeaderboardEntry;
}

export interface Guild {
  id: string;
  name: string;
  description?: string;
  level: number;
  memberCount: number;
  ownerName: string;
  createdAt?: number;
}

export interface GuildsResponse {
  guilds: Guild[];
}

export type GuildDetailsMembersItem = {
  userId: string;
  name: string;
  level: number;
  joinedAt: number;
};

export interface GuildDetails {
  guild: Guild;
  members: GuildDetailsMembersItem[];
}

export interface LotteryEntry {
  userId: string;
  name: string;
  enteredAt: number;
}

export interface LotteryWinner {
  userId: string;
  name: string;
  prize: number;
  wonAt: number;
}

export interface LotteryResponse {
  active: boolean;
  pool: number;
  entryCount: number;
  maxEntries: number;
  entries: LotteryEntry[];
  recentWinners: LotteryWinner[];
}

export interface CommunityStats {
  totalMembers: number;
  activeMissions: number;
  totalCards: number;
  totalGuilds: number;
  totalTransactions: number;
}

export type GetAllCardsParams = {
  tier?: string;
  series?: string;
};

export type GetLeaderboardParams = {
  limit?: number;
};

export type GetGuildsParams = {
  search?: string;
};