# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Shadow Garden API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/auth/otp/send:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP via WhatsApp bot DM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpBody"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendOtpResponse"
        "400":
          description: Invalid phone number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not registered with the bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/otp/verify:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP code and return session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpBody"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/cards:
    get:
      operationId: getAllCards
      tags: [cards]
      summary: Get all cards with optional filtering
      parameters:
        - in: query
          name: tier
          schema:
            type: string
        - in: query
          name: series
          schema:
            type: string
      responses:
        "200":
          description: Cards list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardsResponse"
  /v1/cards/my:
    get:
      operationId: getMyCards
      tags: [cards]
      summary: Get authenticated user's card collection
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User's cards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCardsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/cards/wishlist:
    post:
      operationId: addCardToWishlist
      tags: [cards]
      summary: Notify card owner that you want to trade
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistBody"
      responses:
        "200":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    SendOtpBody:
      type: object
      properties:
        phone:
          type: string
          description: Phone number in international format (e.g. +2347012345678)
      required:
        - phone
    SendOtpResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    VerifyOtpBody:
      type: object
      properties:
        phone:
          type: string
        code:
          type: string
      required:
        - phone
        - code
    AuthSession:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - success
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        level:
          type: integer
        xp:
          type: integer
        balance:
          type: integer
        bank:
          type: integer
        premium:
          type: integer
        bio:
          type: string
        registeredAt:
          type: integer
      required:
        - id
        - name
        - level
        - xp
        - balance
        - bank
...
    Card:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        tier:
          type: string
        series:
          type: string
        description:
          type: string
        attack:
          type: integer
        defense:
          type: integer
        speed:
          type: integer
        totalCopies:
          type: integer
        ownerName:
          type: string
        ownerId:
          type: string
      required:
        - id
        - name
        - tier
        - series
        - attack
        - defense
        - speed
        - totalCopies
    CardsResponse:
      type: object
      properties:
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Card"
        total:
          type: integer
      required:
        - cards
        - total
    UserCard:
      type: object
      properties:
        userCardId:
          type: integer
        card:
          $ref: "#/components/schemas/Card"
        obtainedAt:
          type: integer
      required:
        - userCardId
        - card
        - obtainedAt
    UserCardsResponse:
      type: object
      properties:
        cards:
          type: array
          items:
            $ref: "#/components/schemas/UserCard"
        total:
          type: integer
      required:
        - cards
        - total
    WishlistBody:
      type: object
      properties:
        cardId:
          type: string
      required:
        - cardId
...