# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shadow Garden API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Send OTP via WhatsApp bot DM
 */
export const SendOtpBody = zod.object({
  phone: zod
    .string()
    .describe("Phone number in international format (e.g. +2347012345678)"),
});

export const SendOtpResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Verify OTP code and return session token
 */
export const VerifyOtpBody = zod.object({
  phone: zod.string(),
  code: zod.string(),
});

export const VerifyOtpResponse = zod.object({
  success: zod.boolean(),
  token: zod.string(),
  user: zod.object({
    id: zod.string(),
    name: zod.string(),
    phone: zod.string().optional(),
    level: zod.number(),
    xp: zod.number(),
    balance: zod.number(),
    bank: zod.number(),
    premium: zod.number().optional(),
    bio: zod.string().optional(),
    registeredAt: zod.number().optional(),
  }),
});

/**
 * @summary Get current user full profile and stats
 */
export const GetUserStatsResponse = zod.object({
  profile: zod.object({
    id: zod.string(),
    name: zod.string(),
    phone: zod.string().optional(),
    level: zod.number(),
    xp: zod.number(),
    balance: zod.number(),
    bank: zod.number(),
    premium: zod.number().optional(),
    bio: zod.string().optional(),
    registeredAt: zod.number().optional(),
  }),
  rpg: zod
    .object({
      class: zod.string(),
      hp: zod.number(),
      maxHp: zod.number(),
      attack: zod.number(),
      defense: zod.number(),
      speed: zod.number(),
      dungeonFloor: zod.number(),
    })
    .optional(),
  guild: zod
    .object({
      id: zod.string(),
      name: zod.string(),
      level: zod.number(),
    })
    .optional(),
  rank: zod.number(),
  totalUsers: zod.number(),
  xpNeeded: zod.number(),
});

/**
 * @summary Get current user inventory
 */
export const GetUserInventoryResponse = zod.object({
  items: zod.array(
    zod.object({
      item: zod.string(),
      quantity: zod.number(),
      category: zod.string(),
    }),
  ),
});

/**
 * @summary Get current user achievements and badges
 */
export const GetUserAchievementsResponse = zod.object({
  achievements: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      description: zod.string(),
      icon: zod.string(),
      earnedAt: zod.number(),
    }),
  ),
});

/**
 * @summary Get all shop items by category
 */
export const GetShopItemsResponse = zod.object({
  categories: zod.array(
    zod.object({
      name: zod.string(),
      items: zod.array(
        zod.object({
          id: zod.number(),
          name: zod.string(),
          description: zod.string(),
          price: zod.number(),
          effect: zod.string(),
          category: zod.string(),
        }),
      ),
    }),
  ),
});

/**
 * @summary Buy an item from the shop
 */
export const buyShopItemBodyQuantityDefault = 1;

export const BuyShopItemBody = zod.object({
  itemId: zod.number(),
  quantity: zod.number().default(buyShopItemBodyQuantityDefault),
});

export const BuyShopItemResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
  newBalance: zod.number(),
});

/**
 * @summary Get all cards with optional filtering
 */
export const GetAllCardsQueryParams = zod.object({
  tier: zod.coerce.string().optional(),
  series: zod.coerce.string().optional(),
});

export const GetAllCardsResponse = zod.object({
  cards: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      tier: zod.string(),
      series: zod.string(),
      description: zod.string().optional(),
      attack: zod.number(),
      defense: zod.number(),
      speed: zod.number(),
      totalCopies: zod.number(),
      ownerName: zod.string().optional(),
      ownerId: zod.string().optional(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get authenticated user's card collection
 */
export const GetMyCardsResponse = zod.object({
  cards: zod.array(
    zod.object({
      userCardId: zod.number(),
      card: zod.object({
        id: zod.string(),
        name: zod.string(),
        tier: zod.string(),
        series: zod.string(),
        description: zod.string().optional(),
        attack: zod.number(),
        defense: zod.number(),
        speed: zod.number(),
        totalCopies: zod.number(),
        ownerName: zod.string().optional(),
        ownerId: zod.string().optional(),
      }),
      obtainedAt: zod.number(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Notify card owner that you want to trade
 */
export const AddCardToWishlistBody = zod.object({
  cardId: zod.string(),
});

export const AddCardToWishlistResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get global XP/level leaderboard
 */
export const getLeaderboardQueryLimitDefault = 10;

export const GetLeaderboardQueryParams = zod.object({
  limit: zod.coerce.number().default(getLeaderboardQueryLimitDefault),
});

export const GetLeaderboardResponse = zod.object({
  entries: zod.array(
    zod.object({
      rank: zod.number(),
      userId: zod.string(),
      name: zod.string(),
      level: zod.number(),
      xp: zod.number(),
      guildName: zod.string().nullish(),
    }),
  ),
});

/**
 * @summary Get current user's rank position
 */
export const GetMyRankResponse = zod.object({
  rank: zod.number(),
  total: zod.number(),
  entry: zod.object({
    rank: zod.number(),
    userId: zod.string(),
    name: zod.string(),
    level: zod.number(),
    xp: zod.number(),
    guildName: zod.string().nullish(),
  }),
});

/**
 * @summary Get all guilds
 */
export const GetGuildsQueryParams = zod.object({
  search: zod.coerce.string().optional(),
});

export const GetGuildsResponse = zod.object({
  guilds: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      description: zod.string().optional(),
      level: zod.number(),
      memberCount: zod.number(),
      ownerName: zod.string(),
      createdAt: zod.number().optional(),
    }),
  ),
});

/**
 * @summary Get a specific guild by ID
 */
export const GetGuildParams = zod.object({
  guildId: zod.coerce.string(),
});

export const GetGuildResponse = zod.object({
  guild: zod.object({
    id: zod.string(),
    name: zod.string(),
    description: zod.string().optional(),
    level: zod.number(),
    memberCount: zod.number(),
    ownerName: zod.string(),
    createdAt: zod.number().optional(),
  }),
  members: zod.array(
    zod.object({
      userId: zod.string(),
      name: zod.string(),
      level: zod.number(),
      joinedAt: zod.number(),
    }),
  ),
});

/**
 * @summary Get current lottery state, entries, and recent winners
 */
export const GetLotteryStateResponse = zod.object({
  active: zod.boolean(),
  pool: zod.number(),
  entryCount: zod.number(),
  maxEntries: zod.number(),
  entries: zod.array(
    zod.object({
      userId: zod.string(),
      name: zod.string(),
      enteredAt: zod.number(),
    }),
  ),
  recentWinners: zod.array(
    zod.object({
      userId: zod.string(),
      name: zod.string(),
      prize: zod.number(),
      wonAt: zod.number(),
    }),
  ),
});

/**
 * @summary Get community-wide stats (total members, active missions, etc.)
 */
export const GetCommunityStatsResponse = zod.object({
  totalMembers: zod.number(),
  activeMissions: zod.number(),
  totalCards: zod.number(),
  totalGuilds: zod.number(),
  totalTransactions: zod.number(),
});